refactor(kill)
move metadata and monsters to their own files
