+++ src/commands/rs/_kill.ts
@@ -1,7 +1,7 @@
 import { CommandContext } from "@types-local/commands";
 import { Monster, Monsters, Util } from "oldschooljs";
 import { parseLoot } from "./_rs_utils.js";
-import { CustomMonsters } from "./monsters/index.js";
+import { NewMonsters } from "./monsters/index.js";
 import {
   ActionRowBuilder,
   ButtonBuilder,
@@ -11,60 +11,11 @@ import {
   MessageFlags,
   User,
 } from "discord.js";
-
-const NativeMonsters = [...Monsters].map((m) => m[1]);
-const NewMonsters = [...NativeMonsters, ...CustomMonsters];
+import { metadata } from "./monsters/_kill-metadata.js";
 
 function getInMemoryPartyKey(userId: string) {
   return `${userId}:parties`;
 }
-
-type MonsterMetaData = {
-  [monsterId: number]: {
-    teamBoss: boolean;
-    partySizes: number[];
-    cooldowns: number[];
-  };
-};
-
-// partySizes and cooldowns must have the same length
-const metadata: MonsterMetaData = {
-  13447: {
-    teamBoss: true,
-    partySizes: [1, 2, 3, 4, 5],
-    cooldowns: [120 * 60, 30 * 60, 20 * 60, 15 * 60, 10 * 60],
-  },
-  14176: {
-    teamBoss: true,
-    partySizes: [1, 2],
-    cooldowns: [6 * 60, 3 * 60],
-  },
-  319: {
-    teamBoss: true,
-    partySizes: [1, 2, 3],
-    cooldowns: [1 * 60, 30, 15],
-  },
-  12192: {
-    teamBoss: false,
-    partySizes: [1],
-    cooldowns: [5 * 60],
-  },
-  12215: {
-    teamBoss: false,
-    partySizes: [1],
-    cooldowns: [5 * 60],
-  },
-  12224: {
-    teamBoss: false,
-    partySizes: [1],
-    cooldowns: [5 * 60],
-  },
-  12205: {
-    teamBoss: false,
-    partySizes: [1],
-    cooldowns: [5 * 60],
-  },
-};
 
 function renderActiveParty(
   partyLead: User,
