+++ src/commands/rs/monsters/index.ts
@@ -1,5 +1,8 @@
+import { Monsters } from "oldschooljs";
 import { Araxxor } from "./_araxxor.js";
 import { Nex } from "./_nex.js";
 import { Yama } from "./_yama.js";
 
-export const CustomMonsters = [Araxxor, Nex, Yama];
+const CustomMonsters = [Araxxor, Nex, Yama];
+const NativeMonsters = [...Monsters].map((m) => m[1]);
+export const NewMonsters = [...NativeMonsters, ...CustomMonsters];
